refactor: consolidate AGUI handlers into events.py
Merge handlers.py + workflow_handlers.py into a single events.py file,
following the Slack interface pattern where all event handlers live in
one file.

Key improvements:
- Extract _close_text_if_open(), _ensure_reasoning_open(), _close_reasoning()
  to eliminate 4x duplicated patterns
- Split _finalize_run() into _emit_paused_tools() + _emit_final_state()
- Use _event_value() consistently (was 3 different extraction patterns)
- Add module docstring explaining key concepts
- Remove dead code (unreachable branches, unnecessary try/except)

Before: handlers.py (488) + workflow_handlers.py (143) = 631 lines
After: events.py = 530 lines (-16%)
